added employee CRUD

+++ src/main/java/com/contacts/backend/repositories/EmploeeRepository.java
@@ -3,5 +3,5 @@ package com.contacts.backend.repositories;
 import com.contacts.backend.entity.EmployeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmploeeRepository extends JpaRepository<EmployeeEntity, Long> {
+public interface EmploeeRepository extends JpaRepository<EmployeeEntity, Integer> {
 }
